Cache player id resolve
Resolving the passed in player ID is a crucial command used by many
other commands as well. Most users do not pass in a CFToolsId directly,
resulting in numerous times that the passed ID needs to be resolved
before the requested action can be done.

However, the resolve endpoint is under heavy rate limit (20 requests per
minute), resolving the same and the same ID again every time a new
action with an already resolved ID is a waste of resources. After all,
the resolved IDs are static anyway, there is no known way of when the
result of requesting the same ID (Steam ID, BE GUID or BI UID) might
be diffrent.

This commit will bring resolving the IDs under the cache configuration
umbrella. The default caching time for this function, as they are
static, is very high, so that one ID usually does not need to be
resolved twice. It depends on the caching storage, if it can or will
hold the cached results in the cache, though. Also, enabling the cache
is crucial for this to work as well, which is opt-in by the user of this
sdk.

Discovered in https://github.com/FlorianSW/server-donation-tool/issues/53